% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportInterpolated.R
\name{is.transportInterpolated}
\alias{is.transportInterpolated}
\title{Check validity of interpolated g-computation result object}
\usage{
is.transportInterpolated(x)
}
\arguments{
\item{x}{Result object from \code{transportInterpolated} function}
}
\value{
A boolean indicating whether all components of \code{transportInterpolated} object have the correct types.
}
\description{
A simple helper function that validates whether the components of the given \code{transportInterpolated} object are of the correct types.
}
