% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportTADA.R
\name{is.transportTADA}
\alias{is.transportTADA}
\title{Check validity of TADA result object}
\usage{
is.transportTADA(transportTADAResult)
}
\arguments{
\item{transportTADAResult}{Result object from \code{transportTADA} function}
}
\value{
A boolean indicating whether all components of \code{transportTADA} object have the correct types.
}
\description{
A simple helper function that validates whether the components of the given \code{transportTADA} object are of the correct types.
}
