% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportInterpolated.R
\name{plot.transportInterpolated}
\alias{plot.transportInterpolated}
\title{Visually represent results of transportability analysis using interpolated g-computation}
\usage{
\method{plot}{transportInterpolated}(x, ...)
}
\arguments{
\item{x}{Result from \code{transportInterpolated} function}

\item{...}{Further arguments from previous function or to pass to next function}
}
\value{
A \code{ggplot} object showing the estimates and confidence intervals of the transported effect estimate.
}
\description{
This function is a wrapper for \code{modelsummary::modelplot} to plot effect estimates in a transportability analysis using interpolated g-computation. Note that the transported variance estimate is used in this function.
}
