% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportTADA.R
\name{plot.transportTADA}
\alias{plot.transportTADA}
\title{Plot graphs relevant to transportability analysis using TADA}
\usage{
\method{plot}{transportTADA}(
  x,
  type = "propensityHist",
  bins = 50,
  maxWeight = NULL,
  covariates = NULL,
  effectModifiers = NULL,
  ...
)
}
\arguments{
\item{x}{Result from \code{transportTADA} function}

\item{type}{One of \code{"propensityHist", "propensitySMD", "participationHist", "msm"}. \code{Hist} produces mirrored histograms of estimated probability of treatment between treatment groups (for \code{propensity}), or of estimated participation weights between study and target data (for \code{participation}). \code{SMD} produces SMD plots of covariates between treatment groups (for \code{propensity}). \code{msm} produces plots showing confidence intervals for the model coefficients, which should have the correct standard errors.}

\item{bins}{Number of bins for propensity score/participation weight histograms. This is only used for \code{Hist}.}

\item{maxWeight}{Maximum participation weight value, i.g., range, that user wants to show in the \code{participationHist}.}

\item{covariates}{Vector of strings indicating names of covariates in propensity model}

\item{effectModifiers}{Vector of strings indicating names of effect modifiers in participation model}

\item{...}{Further arguments from previous function or to pass to next function}
}
\value{
A \code{ggplot} object which contains the desired plot.
}
\description{
Plot graphs for assessment of covariate balance and results in a TADA analysis. This function currently supports mirrored histograms, SMD plots and model coefficient plots.
}
