% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportInterpolated.R
\name{summary.transportInterpolated}
\alias{summary.transportInterpolated}
\alias{print.summary.transportInterpolated}
\title{Summarize results of transportability analysis using interpolated g-computation}
\usage{
\method{summary}{transportInterpolated}(object, ...)

\method{print}{summary.transportInterpolated}(x, out = stdout(), ...)
}
\arguments{
\item{object}{Result from \code{transportInterpolated} function}

\item{...}{Further arguments from previous function or to pass to next function}

\item{x}{\code{summary.transportInterpolated} object}

\item{out}{Output stream}
}
\value{
Returns a \code{summary.transportInterpolated} object containing the following components:
\itemize{
\item \code{effect}, \code{link}, \code{mainTreatmentEffect} and \code{mainSE} as in the \code{transportInterpolated} object
\item \code{se}: Estimated standard error of the effect estimate
\item \code{subgroupEffects}: Subgroup effect estimates and their estimated standard errors in the original study, organized into a data frame with specific effect modifier and subgroup information
\item \code{aggregateTargetData}: Summary statistics of effect modifiers in target data
}
}
\description{
Returns summary object which contains the transported effect estimate, organizes subgroup effect estimates into a data frame and produces summary statistics of effect modifiers in the target data, if not already provided.
}
