% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportTADA.R
\name{summary.transportTADA}
\alias{summary.transportTADA}
\alias{print.summary.transportTADA}
\title{Summarize results of a fitted MSM using the TADA approach}
\usage{
\method{summary}{transportTADA}(object, covariates = NULL, effectModifiers = NULL, ...)

\method{print}{summary.transportTADA}(x, out = stdout(), ...)
}
\arguments{
\item{object}{Result from \code{transportTADA} function}

\item{covariates}{Vector of strings indicating names of covariates in propensity model}

\item{effectModifiers}{Vector of strings indicating names of effect modifiers in participation model. When it is \code{NULL}, the \code{prePostTable} shows all the results}

\item{...}{Further arguments from previous function or to pass to next function}

\item{x}{\code{summary.transportTADA} object.}

\item{out}{Output stream.}
}
\value{
The \code{summary.transportTADA} function returns a \code{summary.transportTADA} object containing the following components:
\itemize{
\item \code{propensitySMD}: Table of unweighted and weighted SMDs of covariates between treatment groups. Only propensity weights are used.
\item \code{prePostTable}: Table of unweighted and weighted aggregate level summary as well as pre- and post- difference values of effect modifiers of interest between study data and target data.
\item \code{msmSummary}: Summary object of model object for MSM. The correct variance estimators are set here for \code{glm}, whereas they are set in \code{transportTADA} for \code{survreg} and \code{coxph}.
}
}
\description{
Returns summary object which contains a summary of the fitted MSM, pre- and post-weighting standardized mean differences (SMDs) and pre-post matching table including aggregate level summary of effect modifiers of interest.
}
