% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportGC.R
\name{transportGC}
\alias{transportGC}
\title{Transportability analysis using g-computation}
\usage{
transportGC(
  effectType = c("meanDiff", "rr", "or", "hr"),
  preparedModel,
  targetData,
  bootstrapNum = 500
)
}
\arguments{
\item{effectType}{Type of effect desired for the ATE: \code{"meanDiff"} for mean difference, \code{"rr"} for relative risk, \code{"or"} for odds ratio, and \code{"hr"} for hazard ratio.}

\item{preparedModel}{A \code{transportGCPreparedModel} object. This is obtained by using the \code{transportGCPreparedModel} function to fit an outcome model using the study data.}

\item{targetData}{A target dataset.}

\item{bootstrapNum}{Number of bootstrap datasets to simulate to obtain robust variance estimators}
}
\value{
A \code{transportGC} object with the following components:
\itemize{
\item \code{effect}: Calculated ATE
\item \code{effectType}: Type of effect calculated
\item \code{var}: Estimated variance of ATE estimator, calculated using bootstrap
\item \code{preparedModel}: The \code{transportGCPreparedModel} object used to estimate the ATE
\item \code{bootstrapNum}: Integer indicating number of bootstrap datasets simulated to calculate robust variance estimators.
}
}
\description{
Estimates the coefficients of a marginal structural model (MSM) using g-computation in a generalizability or transportability analysis. In particular, the estimators should be unbiased for the coefficients in the superpopulation or the target population, respectively.
}
\details{
The expected workflow is as follows:

\enumerate{
\item{A researcher who wants to perform a generalizability/transportability analysis collects data from the target population.}
\item{They then request the owner of the study data from which they wish to generalize/transport to provide an outcome model fitted using the study data.}
\item{The owner of the study data runs the \code{transportGCPreparedModel} function on the study data to obtain a \code{transportGCPreparedModel} object which contains the fitted outcome model.}
\item{The owner of the study data provides the \code{transportGCPreparedModel} object to the researcher, perhaps in a \code{.rds} file.}
\item{The researcher uses this function and the provided \code{transportGCPreparedModel} object to perform the analysis using g-computation.}
}

Since model-fitting objects in \code{R} often contain the data used to fit the model, the \code{transportGCPreparedModel} function wipes this data in the model-fitting object and keeps additional information about the name of the response variable, the name of the treatment variable and the levels of treatment. This is to comply with government regulations regarding access and integration of data sources from different countries.

The MSM-fitting functions do not provide correct standard errors as-is. Bootstrap is used to calculate robust variance estimates of the MSM coefficient estimators. Note that these standard errors are only valid conditional on the observed study data because it is not possible to resample the study data when access to it is restricted.
}
