% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportGCPreparedModel.R
\name{transportGCPreparedModel}
\alias{transportGCPreparedModel}
\title{Prepare an outcome model object for \code{transportGC}}
\usage{
transportGCPreparedModel(
  outcomeModel,
  response = NULL,
  responseLevels = NULL,
  treatment,
  treatmentLevels = NULL,
  family = stats::gaussian,
  method = c("logistic", "probit", "loglog", "cloglog", "cauchit"),
  studyData = NULL,
  wipe = T,
  formula = NULL
)
}
\arguments{
\item{outcomeModel}{Either a formula or a \code{glm}, \code{survreg}, \code{coxph}, or \code{polr} object representing the outcome model. Please set \code{model = T} if providing a fitted model object.}

\item{response}{String indicating name of response variable. If \code{NULL}, it will be auto-detected from \code{outcomeModel}.}

\item{responseLevels}{For ordinal responses, vector of strings indicating levels of response variable in the study data. If \code{NULL} and \code{polr} is used, it will be auto-detected using \code{response} and \code{studyData}.}

\item{treatment}{String indicating name of treatment variable. This argument is required.}

\item{treatmentLevels}{Vector of strings indicating levels of treatment variable in the study data. If \code{NULL}, it will be auto-detected using \code{treatment} and \code{studyData}.}

\item{family}{Either a family function as used for \code{glm}, or one of \code{c("coxph", "survreg")}. Only required if \code{outcomeModel} is a formula.}

\item{method}{Link function used for \code{polr}, one of \code{c("logistic", "probit", "loglog", "cloglog", "cauchit")}. Only required if \code{outcomeModel} is a formula and \code{polr} is used.}

\item{studyData}{Data frame of the study data.}

\item{wipe}{Logical indicating whether original study data should be wiped from outcome model-fitting object.}

\item{formula}{The formula used to fit outcomeModel, if outcomeModel is provided as a fitted model object. This is necessary to provide only when using \code{coxph}, \code{survreg} and \code{polr} because these objects do not have the formula components.}
}
\value{
A \code{transportGCPreparedModel} object containing the following components:
\itemize{
\item \code{outcomeModel}: The fitted outcome model with all components containing the data used to fit the model removed
\item \code{response}: String indicating name of response variable
\item \code{treatment}: String indicating name of treatment variable
\item \code{treatmentLevels}: Vector of strings indicating levels of treatment variable
\item \code{family}: The \code{family} argument provided
\item \code{wipe}: The \code{wipe} argument provided
}
}
\description{
An outcome model needs to be fitted using the study data to perform transportability analysis using g-computation. However, \code{glm} and functions in \code{survival} typically contain the data used to fit the model in their respective result objects. This function provides the option to remove most components containing the data from these result objects to comply with data sharing regulations. The party with sole access to the study data may use only this function and provide the results object (possibly in a .rds file) to others who request it.

Note that for time-to-event outcomes, \code{survreg} should used for the outcome model over \code{coxph} because the latter is not meant to produce predicted survival times. The \code{coxph} option is provided, but we warn that the computation time might be prohibitive.
}
