% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportInterpolated.R
\name{transportInterpolated}
\alias{transportInterpolated}
\title{Transportability analysis using interpolated g-computation}
\usage{
transportInterpolated(
  link = c("identity", "log"),
  effectModifiers,
  mainTreatmentEffect,
  mainSE,
  subgroupTreatmentEffects,
  subgroupSEs,
  corrStructure = NULL,
  studySampleSize,
  aggregateStudyData,
  targetData
)
}
\arguments{
\item{link}{Defaults to \code{"identity"}, which corresponds to absolute treatment effects for continuous responses. The \code{"log"} option accommodates relative treatment effects such as relative risk, odds ratio and hazard ratio.}

\item{effectModifiers}{Vector of strings indicating effect modifiers to adjust for}

\item{mainTreatmentEffect}{Estimate of ATE in original study}

\item{mainSE}{Estimate of standard error of estimator of ATE in original study}

\item{subgroupTreatmentEffects}{Estimates of subgroup ATEs in original study. Please provide subgroup ATEs in the order of effect modifiers listed in \code{effectModifiers}, and provide the ATE of the subgroup whose proportion is provided in \code{summaryAggregateData} first in each pair}

\item{subgroupSEs}{Estimates of standard errors of subgroup ATEs in original study. Please provide SEs in the order of effect modifiers listed in \code{effectModifiers}, and provide the SE of the subgroup whose proportion is provided in \code{summaryAggregateData} first in each pair}

\item{corrStructure}{Correlation structure of dichotomized effect modifiers. If target IPD is provided, this will be estimated from the target data, if user input is omitted. If target aggregate data is provided, this will be specified by the user and default to an independent correlation structure if left unspecified.}

\item{studySampleSize}{Sample size of original study}

\item{aggregateStudyData}{Vector of proportions of dichotomized effect modifiers in study data. Please provide proportions of only one category for each effect modifier. This category should correspond to the the first ATE and SE provided for each effect modifier.}

\item{targetData}{May be IPD or aggregate. If aggregate, provide proportions of only one category of dichotomized effect modifiers in a named vector (not a data frame)}
}
\value{
A \code{transportInterpolated} object with the following components:
\itemize{
\item \code{effect}: Transported ATE estimate
\item \code{link}: Denotes the link function used. Absolute treatment effects (i.e. for continuous outcomes) correspond to \code{"identity"}, while relative treatment effects correspond to \code{"log"}
\item \code{var}: Transported variance estimate of effect estimate
\item \code{effectModifiers}: Vector of strings indicating effect modifiers adjusted for
\item \code{mainTreatmentEffect}: Estimate of ATE in original study
\item \code{mainSE}: Standard error of estimator of ATE in original study
\item \code{subgroupTreatmentEffects}: Estimates of subgroup ATEs in original study, as provided
\item \code{subgroupSEs}: Estimates of standard errors of subgroup ATEs in original study, as provided
\item \code{corrStructure}: Correlation structure of effect modifiers used in analysis
\item \code{studySampleSize}: Sample size of original study
\item \code{aggregateStudyData}: Aggregate-level study data, as provided
\item \code{targetData}: Target data, as provided
}
}
\description{
Estimates the average treatment effect (ATE) using interpolated g-computation in a generalizability or transportability analysis. In particular, the estimators should be unbiased for the ATE in the superpopulation or the target population, respectively.
}
\details{
This function transports the ATE estimate from the original study data to the target data by utilizing subgroup effect estimates in the same way as network meta-interpolation \insertCite{harari}{TransportHealth}. As standard errors are transported manually, no bootstrapping is done, unlike other methods supported by \code{TransportHealthR}.
}
\references{
\insertAllCited{}
}
