% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportTADA.R
\name{transportTADA}
\alias{transportTADA}
\title{TransportTADA}
\usage{
transportTADA(
  msmFormula,
  propensityScoreModel = NULL,
  matchingCovariates = NULL,
  propensityWeights = NULL,
  participationWeights = NULL,
  treatment = NULL,
  response = NULL,
  family = stats::gaussian,
  studyData,
  aggregateTargetData
)
}
\arguments{
\item{msmFormula}{A formula for the MSM to be fitted, which usually includes the outcome, the treatment and any effect modifiers.}

\item{propensityScoreModel}{Either a formula or a \code{glm} object representing the model for treatment assignment given covariates.}

\item{matchingCovariates}{A vector of user-specified covariates which are used for the data matching to obtain participation weights when no custom weights provided.}

\item{propensityWeights}{Vector of custom weights balancing covariates between treatments. Providing them will override the formula or model provided by \code{propensityScoreModel}. This vector should have as any entries as the sample size of the study data.}

\item{participationWeights}{Vector of custom weights balancing effect modifiers between study and target populations. This vector should have as any entries as the sample size of the study data.}

\item{treatment}{String indicating name of treatment variable. If \code{NULL}, it will be auto-detected from \code{propensityScoreModel} if provided; otherwise it will remain \code{NULL}. Note that when using custom weights, \code{treatment} should be provided so that \code{summary.transportTADA} and \code{plot.transportTADA} works.}

\item{response}{String indicating name of response variable. If \code{NULL}, it will be auto-detected form \code{msmFormula}.}

\item{family}{Either a \code{family} function as used for \code{glm} such as stats::gaussian(), or one of \code{c("coxph", "survreg")}.}

\item{studyData}{The individual patient-level data (IPD) of study population. Ensure that: 1. All binary variables to be used in the matching should be coded 1 and 0;}

\item{aggregateTargetData}{The aggregate-level data (AgD) of target population. Ensure that: 1. Name columns of mean of continuous variables or proportion of binary variable baselines exactly the same as the column names in the study (IPD) data; 2. Only continuous variables are allowed to consider matching standard deviation (SD) and name the SD column as "variable_sd" in the aggregateTargetData.}
}
\value{
A \code{transportTADA} object containing the following components:
\itemize{
\item \code{msm}: Raw model fit object for MSM of class \code{glm}, \code{survreg} and \code{coxph}, with the correct variance estimators appropriately replaced. If of class \code{glm}, it will have an extra \code{var} component containing the correct variance estimates.
\item \code{propensityScoreModel}: Model of treatment assignment, \code{NULL} if not provided and custom propensity weights are used.
\item \code{propensityWeights}: Propensity weights used. When not \code{NULL}, it will be used as custom inputs from user. When \code{NULL}, it will be obtained by logistic regression.
\item \code{participationWeights}: Participation Weights used. When not \code{NULL}, it will be used as custom inputs from user. When \code{NULL}, it will be obtained by Method of Moments -based method.
\item \code{finalWeights}: Weights used to fit MSM.
\item \code{customPropensity}: Boolean indicating whether custom propensity weights are used.
\item \code{customParticipation}: Boolean indicating whether custom participation weights are used.
\item \code{treatment}: String indicating variable name of treatment.
\item \code{response}: String indicating variable name of response.
\item \code{studyData}: The original individual patient-level data (IPD) of study population.
\item \code{aggregateTargetData}: The aggregate-level data (AgD) of target population.
\item \code{centeredStudyData}: The centered study data for obtaining participation weights with MoM by MAIC::estimate_weights function.
}
}
\description{
Estimates the coefficients of a marginal structural model (MSM) using Target Aggregate Data Adjustment (TADA) Transportability Analysis in a transportability analysis.
}
