% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trimming specification object}
\usage{
trim(value)
}
\arguments{
\item{value}{Threshold for trimming, which should be between 0 and 0.5.}
}
\value{
A \code{trim} object that can be processed by \code{transportIP} and \code{transportTADA} for weight trimming.
}
\description{
Specify trimming parameters for \code{transportIP} and \code{transportTADA}.
}
\details{
Here, trimming refers to excluding observations with extreme weights from the analysis. This package only supports Crump trimming (cite), which trims based on propensity scores rather than weights. Thus, trimming is not supported for weighting components that do not have a well-defined propensity score (\code{final} in \code{transportIP}; \code{participation} and \code{final} in \code{transportTADA}).
}
