% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc.R
\name{trunc}
\alias{trunc}
\title{Truncation specification object}
\usage{
trunc(type = c("quantile", "raw"), value)
}
\arguments{
\item{type}{One of \code{"quantile"} or \code{"raw"}. If \code{"quantile"}, the truncation threshold is the quantile of the weights specified by \code{value}. If \code{"raw"}, the truncation threshold is \code{value}.}

\item{value}{Value of threshold as described above. When \code{type = "quantile"}, this argument should be between 0 and 1.}
}
\value{
A \code{trunc} object that can be processed by \code{transportIP} and \code{transportTADA} for weight truncation.
}
\description{
Specify truncation parameters for \code{transportIP} and \code{transportTADA}.
}
\details{
Here, truncation refers to the procedure of setting weights above a threshold equal to the threshold. Thus, observations with extreme weights are included in the analysis, albeit with reduced weights.
}
